requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
ta==0.11.0
vectorbt==0.26.2

# Market Data
kiteconnect==4.3.0
yfinance==0.2.35
requests==2.31.0
beautifulsoup4==4.12.3

# Backtesting
backtrader==1.9.78.123

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.1
schedule==1.2.1
pytz==2024.1
python-dateutil==2.8.2

# Security
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
